requirements.txt
discord.py==2.1.0
pyautogui==0.9.53
pillow==9.1.1
